176 - Move Comparison view to Pages folder and rename one folder
