+++ frontend/src/components/comparisonTable/Body.jsx
