+++ frontend/src/components/comparisonTable/ComparisonCheckbox.jsx
