+++ frontend/src/components/comparisonTable/Metadata3Table.jsx
