+++ frontend/src/components/comparisonTable/ParentBuildComparison.jsx
