+++ frontend/src/components/comparisonTable/ParentTable.jsx
