+++ frontend/src/components/comparisonTable/Row.jsx
