+++ frontend/src/components/comparisonTable/Status.jsx
